Remove dependency on private webfactory tools and composer packages
webfactory's deployment tool phlough was used during the initial implementation to save time on setting up the server but for open source collaboration it is not a grave show-stopper. See this link to learn more about phlough: https://webfactory.atlassian.net/wiki/spaces/paketdoku/pages/6193506/Phlough

Up until now I was using two private webfactory composer packages that offered some defaults for symfony, doctrine and logging.
This configuration now lives exclusively in the configuration yml files in the AppBundle.

The PackageVersion Entity is no longer final because it caused an error when creating a proxy object of it.