Move project/package lists and "add project" to settings page (Case 162267)
For a cleaner main page, I moved the full list of registered projects and packages (as well as the button to add a new repository) to a separate page rather than displaying it directly on the home page.

This information is rarely used, and we can save the database requests and DOM nodes involved with these ever-growing lists most of the time.

Also, I think Baton looks much cleaner and less overwhelming when using it this way.