Route to search results on server side (Case 162267)
Until now, search results have been loaded via AJAX. Because of this, Tabs with certain search results wouldn't survive reloads or browser restarts. In order to share search results, we needed an extra "sharable" URL in order to share search results that was linked below. The form itself was not visible on the shared site, so you had to go back to do a new search. The current search couldn't be modified this way, you had to input all values from scratch.

As Symfony can easily process forms submitted via GET, I thought this would technically be the simpler way to go. The form will now be submitted via GET, which gives you a sharable URL by default. It will survive reloads and browser restarts. The Form stays on top of the results page and will be prefilled according to the URL parameters.